+++ src/main.rs
@@ -11,8 +11,9 @@ mod parser;
 fn main() {
     let parser: Parser = Parser::new();
     let ntp_server : &str = parser.eval_default_host();
+    let binding_address : &str = parser.eval_binding_address();
     let mut packet : ntp::NTP = NTP::new();
-    let client: client::Client = Client::new(ntp_server);
+    let client: client::Client = Client::new(ntp_server, binding_address);
 
     packet.set_client_mode();
     parser.evaluate(client);