+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.2] - 2020-01-13
+
+### Added
+
+- Support for alternate bind-address, when creating the network socket.
+- Commandline argument for the bind-address added.
+
+
 ## [v1.1.1] - 2020-01-12
 
 ### Fixed